Update README.md
ICSC 2025 – My Solution
This repo has my solution for the International Computer Science Competition 2025 Qualification Round.
I worked on the problems, wrote my answers, and put everything into a PDF so it’s easy to share.

📄 View the Solution here :

It was a great chance to practice problem-solving under time pressure and think in new ways. Even though it’s not perfect, I learned a lot and I’m proud of the effort.